disk_cache: Support int64_t offset in Simple Cache
This CL changes `offset` type to int64_t from int on Read/WriteData, so
that it can accept 2GB+ file as an input.

This CL implements such large file handling on Simple Cache only.
As for in-memory cache and blockfile, it denies if offset is larger than
int32_t range for now. We will support it in the following cls.

Bug: 391398191
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6654324
Commit-Queue: Eriko Kurimoto <elkurin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1491526}

+++ net/disk_cache/simple/simple_entry_operation.cc
@@ -66,7 +66,7 @@ SimpleEntryOperation SimpleEntryOperation::CloseOperation(
 SimpleEntryOperation SimpleEntryOperation::ReadOperation(
     SimpleEntryImpl* entry,
     int index,
-    int offset,
+    int64_t offset,
     int length,
     net::IOBuffer* buf,
     CompletionOnceCallback callback) {
@@ -79,7 +79,7 @@ SimpleEntryOperation SimpleEntryOperation::ReadOperation(
 SimpleEntryOperation SimpleEntryOperation::WriteOperation(
     SimpleEntryImpl* entry,
     int index,
-    int offset,
+    int64_t offset,
     int length,
     net::IOBuffer* buf,
     bool truncate,
@@ -148,7 +148,7 @@ SimpleEntryOperation SimpleEntryOperation::DoomOperation(
 SimpleEntryOperation::SimpleEntryOperation(SimpleEntryImpl* entry,
                                            net::IOBuffer* buf,
                                            net::CompletionOnceCallback callback,
-                                           int offset,
+                                           int64_t offset,
                                            uint64_t sparse_offset,
                                            int length,
                                            size_t sparse_length,
